+++ poem-openapi/tests/union.rs
@@ -519,3 +519,90 @@ async fn generics() {
     assert_eq!(schema_f32_f64.any_of[0], f32::schema_ref());
     assert_eq!(schema_f32_f64.any_of[1], f64::schema_ref());
 }
+
+#[test]
+fn rename_all() {
+    #[derive(Object, Debug, PartialEq)]
+    struct A {
+        value: i32,
+    }
+
+    #[derive(Object, Debug, PartialEq)]
+    struct B {
+        value: String,
+    }
+
+    #[derive(Union, Debug, PartialEq)]
+    #[oai(discriminator_name = "type", rename_all = "camelCase")]
+    enum MyObj {
+        PutInt(A),
+        PutString(B),
+    }
+
+    let schema = get_meta::<MyObj>();
+
+    assert_eq!(
+        schema,
+        MetaSchema {
+            rust_typename: Some("union::rename_all::MyObj"),
+            ty: "object",
+            discriminator: Some(MetaDiscriminatorObject {
+                property_name: "type",
+                mapping: vec![
+                    (
+                        "putInt".to_string(),
+                        "#/components/schemas/MyObj_A".to_string()
+                    ),
+                    (
+                        "putString".to_string(),
+                        "#/components/schemas/MyObj_B".to_string()
+                    ),
+                ]
+            }),
+            any_of: vec![
+                MetaSchemaRef::Reference("MyObj_A".to_string()),
+                MetaSchemaRef::Reference("MyObj_B".to_string()),
+            ],
+            ..MetaSchema::ANY
+        }
+    );
+
+    assert_eq!(
+        MyObj::parse_from_json(Some(json!({
+            "type": "putInt",
+            "value": 100,
+        })))
+        .unwrap(),
+        MyObj::PutInt(A { value: 100 })
+    );
+
+    assert_eq!(
+        MyObj::PutInt(A { value: 100 }).to_json(),
+        Some(json!({
+            "type": "putInt",
+            "value": 100,
+        }))
+    );
+
+    assert_eq!(
+        MyObj::parse_from_json(Some(json!({
+            "type": "putString",
+            "value": "abc",
+        })))
+        .unwrap(),
+        MyObj::PutString(B {
+            value: "abc".to_string()
+        })
+    );
+
+    assert_eq!(
+        MyObj::PutString(B {
+            value: "abc".to_string()
+        })
+        .to_json(),
+        Some(json!({
+            "type": "putString",
+            "value": "abc",
+        }))
+    );
+}